easypost/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
pydantic-settings>=2.7
google-genai>=1.4.0
Pillow>=11.0
python-dotenv>=1.0
python-multipart>=0.0.9
